split AuthContext and AuthProvider
