+++ frontend/src/contexts/useAuth.ts
@@ -1,4 +1,3 @@
-// useAuth.ts
 import { useContext } from "react";
 import { AuthContext } from "@/contexts/AuthContext";
